sql: fix insertion crash
Previously SQL didn't validate for all cases, that the amount of the
source and destination columns during insertion is equal. The problem
was detected when we insert an incorrect amount of values into the
table. For example, the query

insert into t(a) select a, b from t

produced an instance crash. Fixed.

Closes #7132

NO_DOC=bug fix

Co-authored-by: Mergen Imeev <imeevma@gmail.com>