alter: don't swap tuple dictionaries for online space upgrade
We don't rebuild the tuples stored in the space when a space format is
updated, just check that they are compatible with the new format. For
the old tuple fields to be accessible by the new format field names, we
point the old format dictionary to the new format dictionary (a field
dictionary is a reference counted object so the same dictionary can be
used by multiple formats).

This isn't necessary for online space upgrade, because in this case
we upgrade each tuple stored in the space to the new format before
returning it to the user. Moreover, using the new dictionary for the old
tuples would even be harmful in this case, because the old tuples may be
incompatible with the new format, while the space upgrae function may
use the old field names.

Needed for https://github.com/tarantool/tarantool-ee/issues/112

NO_DOC=ee
NO_TEST=ee
NO_CHANGELOG=ee

(cherry picked from commit 73733e51331027424ddce117c09402b7d61db490)