doc: add contributing guide
GitHub suggests to look on the document, when it is present. The banner
is shown for visitors of issues and pull requests pages. We can use this
feature to highlight typical stones on the contributor's road.

My goal is to leave suggestions for typical points of a confusion in a
short document (don't bury into details) and leave links for further
reading.

I hope we should move most of content of the 'open a pull request'
section to a wiki page and possibly merge it with [1] and [2]. It
requires more work and I would consider it as a future step toward good
developer guidelines.

[1]: https://github.com/tarantool/tarantool/wiki/Code-review-procedure
[2]: https://www.tarantool.io/en/doc/latest/dev_guide/developer_guidelines

NO_CHANGELOG=not a user visible change
NO_DOC=nothing requires an update on the website
NO_TEST=it is purely documentation change