applier: make SSLError retryable
A certificate used for an encrypted replication connection may expire,
in which case the user will be forced to reconfigure the connection both
on the master and on the replica to use the new certificate. If the
replica completes the reconfiguration first, it will fail to connect to
the master and never retry, even though the master should complete its
reconfiguration in a moment.

To avoid that, we should make the applier retry to connect on SSLError,
just like it does on SocketError.

Needed for https://github.com/tarantool/tarantool-ee/issues/107

NO_DOC=ee
NO_TEST=ee
NO_CHANGELOG=ee

(cherry picked from commit a1183d6e0116bac4ad6419b892a5736a5355dac6)