replication: retry join automatically
If the error is non-critical, the instance retries retry join
automatically.

Closes #6126

@TarantoolBot document
Title: Retry join automatically (for a timeout)

If the error is non-critical, the instance retries retry join
automatically. For example, if automatic election is used, then a new
instance can find the leader, but while it is sending a join request,
the leader can resign. The join will fail saying the former leader is
read-only. The waiting make sense in this case.