schema: fix sequence_cache_delete
sequence_cache_delete passes a sequence id to mh_i32ptr_del, but
mh_32ptr_del expects a hash key, as returned by mh_i32ptr_find, not
an id. Often, key and id coincide so our tests pass, but this isn't
always a case. Let's fix this.

Note, this isn't a cherry-pick, because the issue was fixed on the
master branch in the scope of commit 4be73c92a646 ("ddl: synchronize
sequence cache with actual data state"), which added a new feature.

Closes #5306