vinyl: drop UPSERT squashing optimization when there is no disk data
The optimization is mostly useless, because it only works if there's no
data on disk. As explained in #5080, it contains a potential bug: if L0
dump is triggered between 'prepare' and 'commit', it will insert a
statement to a sealed vy_mem. Let's drop it.

Part of #5080

NO_DOC=bug fix
NO_CHANGELOG=later