main: free port, random, and memory
These 3 modules are low hanging fruits which right now can be
freed at return from main() without any effort.

There are still a lot of other modules whose freeing is not that
easy. A few hard to untangle knots, and there are more:

- Session, credentials, iproto, and fibers are tied together via
  the latter. Each fiber potentially has a session, its current
  credentials object. Each iproto connection has a session and a
  file descriptor which is stored in the session too.

  The possible solution would be to walk all the fibers and
  destroy them before proceeding to destroy everything else.

- Tuples depend on memtx, and Lua depends on tuples. Because there
  are tuples allocated on memtx->arena. Hence destruction of memtx
  and its arena makes the tuples still stored in Lua invalid.

  It seems Lua should be destroyed first, not last. It would free
  all the refs which might be kept at objects in C in the other
  modules.

- IProto connections leak when iproto is destroyed. They are not
  freed and their descriptors are not closed properly. That
  requires additional preparatory work to destroy them correctly
  on iproto module deconstruction.

Given amount of work, it should be done as a big separate ticket.

Follow up #7259

NO_CHANGELOG=Not a visible change
NO_DOC=Not a visible change
NO_TEST=Not a visible change