session: panic if failed to allocate session
current_session() is called from C code so it must not throw, but it may
if it fails to allocate a session. Practically, this is hardly possible,
because we don't limit the runtime arena, which is used for allocation
of session objects. Still, this looks potentially dangerous.

Gracefully handling an allocation failure in all places where
current_session() may be called would be complicated. Since it's more of
a theoretical issue, let's panic on a session allocation error, like we
do if we fail to allocate other mission critical system objects.

Closes #4735

NO_DOC=code health
NO_TEST=code health
NO_CHANGELOG=code health