replication: retry join automatically
If the error is non-critical, the instance retries join
automatically.

@TarantoolBot document
Title: Retry join automatically (for a timeout)

There are two types of errors: critical and non-critical.
You can recover from non-critical errors. For example, the
connection master turned out to be read-only. It looks like
a configuration error. If the error is non-critical, the
instance retries join automatically. After a critical error
there is no way to recover, because any of these mistakes
are irreparable anyway. For example, vinyl can create some
files. It's not clear what to do with them to try bootstrap
again.

Closes #6126