cmake: use cmake for building libcurl
For unknown reasons in branch 1.10 we are still using autotools for
building libcurl. However, master branch uses CMake for the same
purpose. Using different build systems requires double efforts for
bumping to a new versions of the libcurl. This patch replaces cmake
module by the version from a master branch.

NO_CHANGELOG=It is not a change visible for users.
NO_DOC=It is not a change visible for users.
NO_TEST=Sure, tests would be useful here, but currently we don't test curl
        integration careful and we'll check changes with code review
        only until they are not written.