http client: autodetect system CA certificates
Curl's CMake script looks over well known system CA certificate
locations and choose first existing one to compile the location into the
library. This logic was disabled by a mistake in 2.10.0 (PR #7119). The
patch enables it back.

Without CA certificates the HTTP client will unable to verify server's
certificate, so the only way to perform an HTTPS request would be use
the `verify_peer = false` option -- disable certificate validation at
all.

That's true that build-time autodetection is known source of problems
when a build host is different from host, where the application is run.
Such autodetection is convenient for development, but we generally
disable it to make tarantool executable distributable across hosts with
the same OS (same distro version). It is especially important for
building RPM/Deb packages for our repositories.

However I think that it is okay for detect system CA certificates
location, because it unlikely will vary across hosts with the same OS
distribution version. The alternative is to freeze locations for all
supported OSes, but it is, again, potential source of problems for
community supported packages and for newer OS versions.

NO_DOC=Lack of HTTPS support is definitely broken behavior, there is no
       sense to document it or the opposite.
NO_TEST=A simple test would require to send a request to external host.
        It would not work without internet connection or in a sandbox.
        Such test also would be unstable and would fail from time to
        time due to network conditions. I verified the patch manually. I
        have an idea to add more thorough http client testing later.

Fixes #7372