sql: introduce OpenSpace opcode
Prior to this patch, some opcodes could use a pointer to struct space
that was set during parsing. However, the pointer to struct space is not
really something that defines spaces. A space can be identified by its
ID or name. In most cases, specifying space by pointer works fine, but
truncate() changes the pointer to space, resulting in a sigfault for
prepared statements using the above opcodes. To avoid this problem, a
new opcode has been introduced. This opcode uses the space ID to
determine the pointer to the struct space at runtime and stores it in
the MEM, which is later used in the mentioned opcodes.

Closes #7358

NO_DOC=bugfix