console: rollback txn on eval error if begin was in expr
The issue is if in binary remote console a error is thrown in expression
like "box.begin() error('something') box.commit()" then it is
overwritten by iproto check for active transactions at the end of eval.
The solution is to rollback active transaction in this case in console
code before iproto check.

Let's also assert in tests a behaviour that after successful transaction
begin it stays active in next evaluations (successful or not) until
explicit rollback.

Closes #7288

NO_DOC=minor change