box: fix finding snapshot in upgrade script
The upgrade script first tries to determine if the node is booted from
old snaps not recoverable on current Tarantool versions. If this is the
case, it sets up special triggers so that snaps are automatically
converted to a suitable format.

This happens before box.cfg{}, so the workdir is not set at this point
in time, and the upgrade script should take configured work_dir into
account explicitly. Fix this.

Closes #7232

NO_DOC=bugfix