decimal: add the library into the module API
The existing API was designed to allow a user to allocate decimal values
in his/her own memory (say, on the stack). This commit mostly just
exports the functions to the module API, so it allows a user of the
module API to do the same. As result the decimal number structure is
exposed too (it is NOT opacue). A module API user should know a size and
an alignment of the structure.

The downside is that we can't freely modify the data layout of decimals
in tarantool anymore (without ABI breakage). It is deliberate decision:
I doubt that we'll need to change decimals representation in a future.

Added `STRUCT_DECIMAL` macro to allow a module API user to, say, include
decNumber library and alias tarantool's decimal type to decNumber's
structure.

Kept `strtodec()` private: it does not follow the common `decimal_*()`
naming pattern. We can add it into the module API later with another
name.

Kept `decimal_str()` private: it looks dangerous without at least good
explanation when data in the static buffer are invalidated. There is
`decimal_to_string()` that writes to an explicitly provided buffer.

Replaced `enum` with `#define` for better portability: enum has
implementation defined size and we shoudn't use it, when ABI
compatibility matters.

Changed headers order in test files to alias decimal_t to decNumber and
don't step into the 'conflicting types' error.

Part of #7228

@TarantoolBot document
Title: Module API for decimals

See the declarations in `src/lib/core/decimal.h` in tarantool sources.